+++ login.php
@@ -10,15 +10,20 @@ require_once __DIR__.'/init.php';
 // Include functions
 require_once PATH.INC.'/functions.php';
 
+// Start output buffering
+ob_start();
+
 // Start the session
 session_start();
 
 // Create a Login object
 $rs_login = new Login;
 
-// Check whether the user is logging out and log them out if the session cookie is set
-if(isset($_GET['action']) && $_GET['action'] === 'logout')
+// Check whether the user is logging out
+if(isset($_GET['action']) && $_GET['action'] === 'logout') {
+	// Log the user out if the session cookie is set, otherwise redirect them to the login page
 	isset($_COOKIE['session']) ? $rs_login->userLogout($_COOKIE['session']) : redirect('login.php');
+}
 ?>
 <!DOCTYPE html>
 <html>
@@ -33,14 +38,17 @@ if(isset($_GET['action']) && $_GET['action'] === 'logout')
 	<body class="login">
 		<div class="wrapper">
 			<h1><a href="/"><?php getSetting('site_title'); ?></a></h1>
-			<?php $error = isset($_POST['submit']) ? $rs_login->userLogin($_POST) : ''; echo $error; ?>
+			<?php echo isset($_POST['submit']) ? $rs_login->userLogin($_POST) : ''; ?>
 			<form class="data-form" action="" method="post">
-				<p><label for="username_email">Username or Email<br><input type="text" name="username_email" autofocus></label></p>
+				<p><label for="login">Username or Email<br><input type="text" name="login" autofocus></label></p>
 				<p><label for="password">Password<br><input type="password" name="password"></label></p>
-				<p><label for="captcha">Captcha<br><input type="text" name="captcha"><img id="captcha" src="<?php echo INC.'/captcha.php'; ?>"></label></p>
-				<p><label class="checkbox-label"><input type="checkbox" name="remember_login" value="0"> <span>Keep me logged in</span></label></p>
+				<p><label for="captcha">Captcha<br><input type="text" name="captcha" autocomplete="off"><img id="captcha" src="<?php echo INC.'/captcha.php'; ?>"></label></p>
+				<p><label class="checkbox-label"><input type="checkbox" name="remember_login" value="checked"> <span>Keep me logged in</span></label></p>
 				<input type="submit" class="button" name="submit" value="Log In">
 			</form>
 		</div>
 	</body>
 </html>
+<?php
+// End output buffering
+ob_end_flush();